@@ -1,16 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net8.0-windows</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
+		<TargetFramework>net9.0-windows</TargetFramework>
 		<UseWPF>true</UseWPF>
 		<AllowUnsafeBlocks>True</AllowUnsafeBlocks>
 	</PropertyGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="Microsoft.AspNetCore.Hosting.Server.Abstractions" Version="2.2.0" />
-	  <PackageReference Include="Microsoft.Web.WebView2" Version="1.0.2535.41" />
+	  <PackageReference Include="Microsoft.AspNetCore.Hosting.Server.Abstractions" />
+	  <PackageReference Include="Microsoft.Web.WebView2" />
 	</ItemGroup>
 
 	<ItemGroup>
