@@ -1,4 +1,5 @@
-﻿using DualDrill.CLSL.Language.IR.ShaderAttribute;
+﻿using DualDrill.CLSL.Language.IR.Expression;
+using DualDrill.CLSL.Language.IR.ShaderAttribute;
 using DualDrill.CLSL.Language.Types;
 using System.Collections.Immutable;
 
@@ -7,7 +8,7 @@ namespace DualDrill.CLSL.Language.IR.Declaration;
 public sealed record class ParameterDeclaration(
     string Name,
     IShaderType Type,
-    ImmutableHashSet<ShaderAttribute.IShaderAttribute> Attributes) : IDeclaration
+    ImmutableHashSet<IShaderAttribute> Attributes) : IDeclaration, IVariableIdentifierResolveResult
 {
 }
 
