@@ -4,7 +4,9 @@ public sealed class ShaderRuntimeMethodAttribute : Attribute, IShaderMetadataAtt
 {
 }
 
-public sealed class VecPositionalValueConstructorMethodAttribute : Attribute, IShaderMetadataAttribute { }
-public sealed class VecBroadcastConstructorMethodAttribute : Attribute, IShaderMetadataAttribute { }
-public sealed class VecConversionConstructorMethodAttribute : Attribute, IShaderMetadataAttribute { }
-public sealed class VecZeroConstructorMethodAttribute : Attribute, IShaderMetadataAttribute { }
+public interface IVecConstructorMethodAttribute : IShaderMetadataAttribute { }
+
+public sealed class VecPositionalValueConstructorMethodAttribute : Attribute, IVecConstructorMethodAttribute { }
+public sealed class VecBroadcastConstructorMethodAttribute : Attribute, IVecConstructorMethodAttribute { }
+public sealed class VecConversionConstructorMethodAttribute : Attribute, IVecConstructorMethodAttribute { }
+public sealed class VecZeroConstructorMethodAttribute : Attribute, IVecConstructorMethodAttribute { }
