@@ -88,18 +88,30 @@ sealed class RuntimeDefinitions : ISingleton<RuntimeDefinitions>
                     if (paramTypes.All(runtimeTypes.ContainsKey))
                     {
                         var parameterDecls = paramTypes.Select(t => new ParameterDeclaration(t.Name, runtimeTypes[t], []));
-                        result.Add(m, ShaderFunction.Instance.GetFunction(m.Name, [.. parameterDecls.Select(p => p.Type)]));
+                        var parameterTypes = parameterDecls.Select(p => p.Type).ToArray();
+                        if (m.Name == "vec2" && parameterTypes.Length == 2 && rt is VecType { ElementType: BoolType, Size: N2 })
+                        {
+                            var fl = ShaderFunction.Instance.GetFunction(m.Name, rt, parameterTypes);
+                        }
+                        var f = ShaderFunction.Instance.GetFunction(m.Name, rt, parameterTypes);
+                        result.Add(m, f);
                     }
                 }
             }
         }
 
+        var vec4f32t = ShaderType.GetVecType(N4.Instance, ShaderType.F32);
         foreach (var c in typeof(Vector4).GetConstructors())
         {
             var parameters = c.GetParameters();
-            var m = typeof(DMath).GetMethod("vec4", 0, parameters.Select(p => p.ParameterType).ToArray())
-                    ?? throw new NotSupportedException($"constructor projection not found for {c}");
-            result.Add(c, result[m]);
+            if (!parameters.All(p => runtimeTypes.ContainsKey(p.ParameterType)))
+            {
+                continue;
+            }
+            var f = ShaderFunction.Instance.GetFunction("vec4", vec4f32t, [
+                ..parameters.Select(p => runtimeTypes[p.ParameterType])
+            ]);
+            result.Add(c, f);
         }
 
         return result;
