@@ -1,6 +1,6 @@
 ﻿using DualDrill.Engine.Shader;
-using DualDrill.Graphics;
 using DualDrill.ILSL;
+using System.Reflection;
 namespace DualDrill.Server.Services;
 
 public sealed class ILSLDevelopShaderModuleService
@@ -20,4 +20,8 @@ public sealed class ILSLDevelopShaderModuleService
         [nameof(QuadShader)] = new QuadShader(),
 
     };
+
+    public IReadOnlyDictionary<string, Assembly> KnownAssemblies { get; } = ((IEnumerable<Assembly>)[
+        typeof(MinimumTriangle).Assembly,
+    ]).Select(a => KeyValuePair.Create<string, Assembly>(a.FullName, a)).ToDictionary();
 }
